Break cyclic dependencies
